requirements.txt
requests==2.28.2
pandas==2.0.1
numpy==1.24.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0
nba_api==1.1.9
nfl-data-py==0.3.1
scikit-learn==1.2.2
streamlit==1.22.0
plotly==5.14.1
selenium==4.9.1
webdriver_manager==3.8.6
espn-api==0.8.0
sportsipy==0.6.0
aiohttp==3.8.4
